requirements.txt
Pillow==10.0.0  # Image manipulation library